Don't delete passwords upon editing
Due to an argument change, the pass backend assigned a variable to the
wrong parameter, causing the existing content to be lost between
starting an edit and the dialog box showing up. This patch also contains
some small fixes to deal with edge cases of passwords named after valid
pass commands.