Severely increase queue processing speed
The previous system had a few logic errors, which would cause the main loop
to sleep (to save CPU cycles) too often, such as when we have multiple modules
loaded with one or more having an empty queue, instead of only when there is
really nothing left to do.

Do note that, when an exception occurs in all loaded modules in the same loop,
the system will still sleep shortly. However, this is extremely unlikely and
the kind of situation where we probably want to prevent going through the
queues at full power anyway.