Another huge module queue processing speedup
The search function of ViewManager is extremely slow. By only running it
for every 100 loaded entries or when the queue just got emptied instead of
after every single queue entry we get huge speed improvements.

It is also nonsensical to run search on a module the moment it is binded, as
the list cannot possibly contain any entries then yet. That condition was
removed too.