Add extra .png image for README rendering
Neither GitHub nor NotABug render the SVG in the README. One fix is to use a
special URL, but that ties the repository into GitHub or NotABug and makes the
image a tracker. So instead, I'm creating a small PNG version to use in the
README.

For more information, see https://github.com/isaacs/github/issues/316.